cli: Extract `query::serialize` to crate root
