+++ crates/typst-cli/src/main.rs
@@ -25,9 +25,11 @@ use clap::Parser;
 use clap::error::ErrorKind;
 use codespan_reporting::term;
 use codespan_reporting::term::termcolor::WriteColor;
-use typst::diag::HintedStrResult;
+use ecow::eco_format;
+use serde::Serialize;
+use typst::diag::{HintedStrResult, StrResult};
 
-use crate::args::{CliArguments, Command};
+use crate::args::{CliArguments, Command, SerializationFormat};
 use crate::timings::Timer;
 
 thread_local! {
@@ -103,6 +105,27 @@ fn print_error(msg: &str) -> io::Result<()> {
 
     output.reset()?;
     writeln!(output, ": {msg}")
+}
+
+/// Serialize data to the output format and convert the error to an
+/// [`EcoString`].
+fn serialize(
+    data: &impl Serialize,
+    format: SerializationFormat,
+    pretty: bool,
+) -> StrResult<String> {
+    match format {
+        SerializationFormat::Json => {
+            if pretty {
+                serde_json::to_string_pretty(data).map_err(|e| eco_format!("{e}"))
+            } else {
+                serde_json::to_string(data).map_err(|e| eco_format!("{e}"))
+            }
+        }
+        SerializationFormat::Yaml => {
+            serde_yaml::to_string(data).map_err(|e| eco_format!("{e}"))
+        }
+    }
 }
 
 #[cfg(not(feature = "self-update"))]
